+++ webextension/js/panel/panel-init.js
@@ -16,7 +16,6 @@ async function baseInit() {
 	;
 	const preferences = await getPreferences('mode', 'panel_height', 'panel_width', 'check_enabled');
 	body.classList.toggle('delegated-version', preferences.get('mode') === 'delegated');
-	body.classList.toggle('simple-version', preferences.get('mode') === 'simplified');
 	body.classList.toggle('normal-version', preferences.get('mode') === 'normal');
 
 	const panel_height = (preferences.get('mode') !== 'normal') ? 215 : preferences.get('panel_height'),